Loans cant be made

+++ ForexFun/ClientApp/src/app/dashboard/dashboard.component.ts
@@ -32,7 +32,7 @@ export class DashboardComponent implements OnInit {
       }
     );
     this.updateState();
-    Observable.interval(60000).subscribe((ignore) => {
+    Observable.interval(10000).subscribe((ignore) => {
       this.updateState();
     });
   }
@@ -60,6 +60,7 @@ export class DashboardComponent implements OnInit {
   buyCurrency(currencyId, amount) {
     console.log("Currency: " + currencyId);
     console.log("Amount: " + amount);
+    if (Number(amount) < 0) return;
     var currencyWallet = this.wallets.find(wallet => wallet.CurrencyId === currencyId);
     var usdWallet = this.wallets.find(wallet => wallet.CurrencyId === 1); //TODO
     var exchangeCurrency = this.currencies.find(currency => currency.CurrencyId === currencyId);
@@ -73,13 +74,14 @@ export class DashboardComponent implements OnInit {
       this.http.post(this.url + "api/wallets", newWallet).subscribe((data: any) => {
         if (data.Succeeded == true) {
           console.log('Success currency');
+          this.updateState();
         }
       });
     }
     else {
       currencyWallet.Amount += Number(amount);
       console.log("Putting currency wallet: ");
-      this.http.put(this.url + "api/wallets/" + currencyWallet.Id, currencyWallet).subscribe();
+      this.http.put(this.url + "api/wallets/" + currencyWallet.Id, currencyWallet).subscribe((data) => this.updateState());
     }
   }
 
@@ -87,24 +89,42 @@ export class DashboardComponent implements OnInit {
     console.log("Currency: " + currencyId);
     console.log("Amount: " + amount);
     var exchangeCurrency = this.currencies.find(currency => currency.CurrencyId === currencyId);
-    var usdWallet = this.wallets.find(wallet => wallet.CurrencyId === 1); //TODO
+    var usdWallet = this.wallets.find(wallet => wallet.CurrencyId === 1);
     usdWallet.Amount += Number(amount) * exchangeCurrency.Value*0.99;
     console.log("Putting USD wallet: ");
     this.http.put(this.url + "api/wallets/" + usdWallet.Id, usdWallet).subscribe();
     var currencyWallet = this.wallets.find(wallet => wallet.CurrencyId === currencyId);
+    if (Number(amount) > currencyWallet.Amount) return;
     if (Math.abs(Number(amount) - currencyWallet.Amount) < Number.EPSILON) {
       this.http.delete(this.url + "api/wallets/" + currencyWallet.Id).subscribe((data: any) => {
         if (data.Succeeded == true) {
           console.log('Success delete wallet');
+          this.updateState();
         }
       });
     }
     else {
       currencyWallet.Amount -= Number(amount);
       console.log("Putting currency wallet: ");
-      this.http.put(this.url + "api/wallets/" + currencyWallet.Id, currencyWallet).subscribe();
+      this.http.put(this.url + "api/wallets/" + currencyWallet.Id, currencyWallet).subscribe((data) => this.updateState());
     }
+    this.updateState();
   }
+
+  updateState() {
+    this.http.get<Currency[]>(this.url + 'api/dashboard').subscribe(result => {
+      this.currencies = result;
+      console.log(result);
+    }, error => console.error(error));
+    this.http.get<Wallet[]>(this.url + 'api/wallets/' + this.Id).subscribe(result => {
+      this.wallets = result;
+      this.wallets.forEach((wallet, index) => {
+        this.http.get<Currency>(this.url + 'api/Currencies/' + wallet.CurrencyId).subscribe(resultCurrency => {
+          console.log(wallet.Id);
+          wallet.Name = resultCurrency.Name;
+        }, error => console.error(error))
+      });
+    }, error => console.error(error));
 }
 
 interface Wallet {
@@ -119,22 +139,4 @@ interface Currency {
   CurrencyId: number;
   Name: string;
   Value: number;
-}
-
-  constructor(private http: HttpClient, private userService: UserService) { }
-
-  updateState() {
-    this.http.get<Currency[]>(this.url + 'api/dashboard').subscribe(result => {
-      this.currencies = result;
-      console.log(result);
-    }, error => console.error(error));
-
-    this.http.get<Wallet[]>(this.url + 'api/wallets/' + 1).subscribe(result => {
-      this.wallets = result;
-      this.wallets.forEach((wallet, index) => {
-        this.http.get<Currency>(this.url + 'api/Currencies/' + wallet.CurrencyId).subscribe(resultCurrency => {
-          console.log(wallet.Id);
-          wallet.Name = resultCurrency.Name;
-        }, error => console.error(error))
-      });
-    }, error => console.error(error));
+}
